Adds support for sets.
Flags: -set-names
Metrics: nftables_set_metadata, nftables_set_size

Fixes Dockerfile to make the libcap installation cacheable.

This uncovered a few bugs in google-nftables regarding parsing of
sets. E.g. IsMap and HasTimeout are based on the wrong flags, and
KeyType is never populated with a complete SetDatatype. The "ismap"
and "keytype" labels will be wrong until that's fixed.